simd_array: use power-of-two splitting
A split in N/2 and N-N/2 lead to holes in the data. N-N/2 and N/2 could
still lead to holes at the next recursion level. The most efficient
splitting wrt. to reductions is to have the first member use a power of
2 size. That's implemented now with a simple recursive constexpr
function that determines the next power of 2 from N-N/2.

refs #613

Signed-off-by: Matthias Kretz <kretz@kde.org>