recursive implementation of simd_mask_array
Modified simd_array::operator== (and friends) to return mask_type
instead of bool. This is required to pass the arithmetics unittests.

This required mask_type to work in this new context, so I started on
rewriting it into recursive form, using the same storage_type0/1 split
as simd_array uses.

refs #613

Signed-off-by: Matthias Kretz <kretz@kde.org>