work around ICC madness: invalid SSE intrinsic definitions
ICC happily defines SSE4 intrinsics even if the commandline options
specify that only SSE2 may be used for the target. Since ICC
additionally doesn't treat __m128 as a fundamental type (which it really
should...) Vc failed to compile if <random> (from libstdc++) was
included.

EVIL workaround: define the include guard of the ICC intrinsic headers
to stop the madness. This is only a solution for Vc 0.7. For Vc 1.0 I
want to have a cleaner solution (which I hope, should be easier with
C++11 and perfect forwarding).

fixes #636

Signed-off-by: Matthias Kretz <kretz@kde.org>