types.h: bring the impl/types.h headers in line
Now for each impl the *_v and *_m typedefs are in the types.h header.

Every impl has is_vector and is_mask traits to identify Vector<T> and
Mask<T> types of the specific impl.

Moved VectorAlignment constexpr to types.h as well. While at it I upped
the alignment for Vc::Scalar to 8 instead of 4. 4 may be unsafe on 64bit
systems.

Signed-off-by: Matthias Kretz <kretz@kde.org>