SSE: refactor intrinsics reimplementation
Vc relied on x86intrin.h to declare only the intrinsics that were
actually available. This broke first with MSVC, later with ICC, and now
with GCC. It is too messy to continue to work around it!

The issue is that I want function calls in the Vc implementation that
either are calls to the intrinsics or calls to the reimplementation of
the same functionality. But my approach of declaring functions with a
leading underscore got me in huge trouble.

The fix rename functions that reimplement intrinsics such that they
drop the leading underscore. Thus calls use the function name without
the leading underscore. If the intrinsic is actually available a macro
is used to add the necessary underscore for code in Vc to find the
intrinsic.

refs #656

Signed-off-by: Matthias Kretz <kretz@kde.org>