cast test: split in four executables
The complete cast test is way too huge for the compiler. Compilation
time with GCC can go up to 1h and memory usage is at ~4GB. The number of
possible type casts is extremely large, though. And any of those
conversions can potentially fail, therefore it has to be tested. To make
the test somewhat more manageable I:
* converted the type list arguments to the test functions to type pairs
  of To/From types. Thus, much of the implicit test information (that
  came from a list of destination types internal to the test function)
  is now an explicit test function parameter and visible in the output
  of the test run. This makes the test output a lot more verbose and
  easier to understand.
* split the list of To/From pairs into four parts that are selected via
  a #define. Thus there are four test binaries where there was one
  before. This makes test & development cycles quicker (but still not
  quick enough).