fix AVX simd_cast uint_v -> float_v
rounding was off for input values greater than 2^31. The previous
implementation had to round twice (1: cvt; 2: add) which created a
larger jitter. The fix uses cvt to double because the mantissa of double
stores enough bits for the 32-bit integer to get rid of the rounding on
cvt. As inputs greater than 2^31 are not the norm, the function first
tries to live without the extra complexity.

Signed-off-by: Matthias Kretz <kretz@kde.org>