simdarray binary operators: disable promotion >int
There is no generic support for 64-bit int SIMD types. Therefore
promotion to a 64-bit integral type (including `long` because it can
potentially have 64 bits) also is not done. But if one of the operands
is a scalar type that is larger than int then the operator is disabled
altogether. We do not want an implicit demotion.

Signed-off-by: Matthias Kretz <kretz@kde.org>