simd_cast: improve offset == 0 calls
with offset == 0 and some combination of To/From types the fallback
(scalar copy) implementation was used, even though simply dropping the
offset from the call would improve the implementation selection
strategy. This also fixes a bug in mask conversion. Though the bug is
still latent in the fallback conversion function.

Signed-off-by: Matthias Kretz <kretz@kde.org>