Iterators: work around GCC ICE
GCC may fail (ICE) if it has to apply std::remove_reference to a type
with may_alias attribute. Since Vector::operator[] likely returns such a
type, the original declaration of the iterator value type was asking for
failure. Instead it should be safe and generic enough to just use
V::EntryType as iterator value type.

Signed-off-by: Matthias Kretz <kretz@kde.org>