simdize: implement Adapter constructors
Implemented broadcasts for simdized types. Thus, an object of the scalar
type can be used as the constructor argument to broadcast the data of
this one object to all SIMD entries in the vectorized object. This is
more complex to get right for std::array, which requires double braces
in the constructor.

Signed-off-by: Matthias Kretz <kretz@kde.org>