Common: move MayAlias helper to types.h
The MayAlias alias template is generally useful, and therefore should go
in one of the headers that is included first. common/types.h fits the
bill.
At this point it is not clear whether the MaskBool specialization is
still needed. I commented it out and it continues to work fine. I'll
give it a bit of time and then I'll drop the specialization.

Signed-off-by: Matthias Kretz <kretz@kde.org>