SSE: fix alignr intrinsic interface
clang complained when the template argument to alignr_epi8 did not fit
into the range of uchar. So use size_t instead, since that's what the
caller function has as a type. Now the immediate may be waay off.
Masking the value off at the lower 8 bits yields a valid instruction.
Obviously calls with large s are UB.

Signed-off-by: Matthias Kretz <kretz@kde.org>