SimdArray: Simplify alignment calculation
The alignment of SimdArray should be independent from the underlying
VectorType. Therefore it's sure not a good idea to have VectorSize so
prominent in the alignment calculation. Now it uses the quotient
sizeof(VectorType) / VectorType::size() to calculate the alignment. This
quotient signifies the size of a scalar entry in a way that a difference
to sizeof(T) would be accounted for.

Signed-off-by: Matthias Kretz <kretz@kde.org>