ctest: Fix compiler detection on MacOS X
It's wrong to first look for g++ on Mac, since clang is the default
compiler. We find the default by looking for c++ first. This makes the
g++ version string harder to parse, but hopefully the logic is solid
enough there. The clang version string generated on Mac is pretty bad,
but I think I got the necessary string replacement to extract a proper
clang version string.

Signed-off-by: Matthias Kretz <kretz@kde.org>