Generalize AlignedBase
VectorAlignedBase(T) provided a solution for a simple fix to correct
alignment on heap allocation. But it was specialized for Vector types
needlessly. Now AlignedBase is a general base class for types that need
a specific alignment. VectorAlignedBase and MemoryAlignedBase are
aliases that use a type parameter to automatically derive the alignment
from the type.

Refs gh-35

Signed-off-by: Matthias Kretz <kretz@kde.org>