Assert that subscripting is not broken
On ia32, the alignment of double is 4 Bytes and thus half of
sizeof(double). This can break the logic in the gather/scatter
subscripting implementation as it assumes the sizeof a structure is a
multiple of the vector type (which is only N * 4 for Scalar::double_v).

For now, have the test skip compilation of the gatherStruct
implementation and issue an expected failure.

Refs: gh-44
Signed-off-by: Matthias Kretz <kretz@kde.org>