Don't adjust the base address for MASKMOVDQU
A scatter containing a 0 in indexes may lead to a segfault (or bus
error).  It happens on the _mm_maskmoveu_si128 call which uses the base
address adjusted by -4. I guess this leads to a page fault on a page
that isn't mapped - even though the actual Bytes that are stored go into
a different page (which would be fine).

Refs: gh-86
Signed-off-by: Matthias Kretz <kretz@kde.org>