Use 64-bit maskstores to fix the segfault
Apparently, the 128-bit maskstores lead to a segfault if the base
address is at the end of an allocated memory region. Using 64-bit (MMX)
maskstores solves the issue. Whether it's solved 100% is unclear, the
MASKMOV variant of the interleaving code is certainly not ready for
becoming the default anytime soon.

Fixes: gh-86
Signed-off-by: Matthias Kretz <kretz@kde.org>