SimdArray: make fromOperation smarter
* Atomic Simd(Mask)Array::fromOperation always called
Common::actual_value and thus would unpack atomic Simd(Mask)Array
objects, even if the function to be called was expecting a
Simd(Mask)Array parameter. Now, there's a smart function in between
determining which arguments need a transformation via actual_value.
* ldexp/frexp require breaking up atomic SimdArray: For const-ref
SimdArray, the implementation is simple: just simd_cast it.
But for the pointer parameter that frexp requires we have an aliasing
problem.
* The aliasing issue can be solved with a MayAlias. However, GCC might
ICE if this reinterpret_cast target type is declared with may_alias. If
it doesn't ICE, GCC warns about ignoring the attribute. Therefore, ifdef
the MayAlias so that GCC doesn't see it.
* Transform ldexp/frexp parameters correctly:
** Extend Segment to handle pointer to T
** Extend Segment with a asSimdArray function
** Add actual_value overload that turns a Segment into a SimdArray via
the new asSimdArray function. Note that this function must use an rvalue
reference as otherwise the forwarding reference will grab it.

Refs: gh-75
Signed-off-by: Matthias Kretz <kretz@kde.org>