Reduce SFINAE usage to help ICC understand the code
ICC fails to do SFINAE as in the previous implementation. Change the
code to rely on the first parameter of the call operator to select the
right overload. The return value can simply be passed by reference -
nullptr still works for marking a void return type because everything is
nicely forwarded via forwarding refs and std::forward.

Refs: gh-93
Signed-off-by: Matthias Kretz <kretz@kde.org>