Reduce number of actual_value overloads
Now that the logic is in place to test whether a function likes a
parameter unwrapped via actual_value or not the need for a default
implementation of actual_value that forwards its argument to the return
value is gone. Once this overload is removed a number of overloads can
be dropped since the greedy forwarding reference is gone.

Signed-off-by: Matthias Kretz <kretz@kde.org>