Don't warn about may_alias type arguments to MayAlias<T>
GCC warns about a second application of the `may_alias` attribute to a
type. Thus, MayAlias<MayAlias<T>> would warn. In generic code it is
sometimes important to apply `may_alias` where the modified type might
already have the attribute or not. In such cases we don't care about the
warning.

Signed-off-by: Matthias Kretz <kretz@kde.org>