Simdize: rewrite brace/paren traits for ICC support
ICC still fails to do SFINAE correctly on the single-brace trait. But it
seems to work fine for the parent and double-brace traits. So redo the
Adapter ctor logic to test for paren and double-brace construction
instead of single-brace and double-brace. This leads to paren
construction being preferred over single-brace construction, though.

Signed-off-by: Matthias Kretz <kretz@kde.org>