MIC: gatherImplementation requires a non-Segment type
The SimdArray::gather/scatter implementation would forward a Segment
index argument to the underlying Vector objects. This works for all but
MIC, since they simply use the subscript operators for gather and
scatter. MIC, OTOH requires a SimdArray<int, 8> instead, so ensure that
any Segment index argument is turned into a SimdArray on forward to the
Vector member.

Signed-off-by: Matthias Kretz <kretz@kde.org>