MSVC: work around miscompilation
With MSVC, use a 32-bit integer, since the 64-bit integer miscompiles.

While looking at the SuccessiveEntries class, I decided this should
never be required in the context of a constant expression. What I really
want is the compiler to optimize the abstraction away. So Vc_INTRINSIC
is what I really want.

Signed-off-by: Matthias Kretz <kretz@kde.org>