Tests: fix corner case on simd_for_each
With 99 entries in the vector and iterating from next(begin) to end() or
data.size(), if
a) V::Size == 2
b) alignof(*next(begin)) == memory_alignment<V>
then there's no need for a scalar step, failing the tests expectations.

The simple fix is to have 100 entries in the vector, so that it has to
iterate over an odd number of values, forcing it to always have at least
one scalar step.

Signed-off-by: Matthias Kretz <kretz@kde.org>