Avoid UB on misaligned access and conversions
The is_conversion_undefined test missed the float-float cases. Avoid all
UB (even if it is in creation of the input values) to remove the noise
when running with UBSAN.

Overload loads and stores that didn't care for alignment with aligned
and unaligned variants. If UBSAN complains, changed the unaligned
implementation to use memcpy.

Fixes loadstore failures with Clang.

Signed-off-by: Matthias Kretz <kretz@kde.org>