MSVC: work around warning C4396
For some reason Microsoft reads from the standard that the combination
of

1. friend
2. function template specialization
3. inline

is not allowed. Can't find this anywhere in the standard (there's a
paragraph that comes close, which is about explicit template
instantiation (14.7.2 p1)).

Anyway, this can be worked around by using a broader friend declaration.
I.e. befriend the complete function template.

Signed-off-by: Matthias Kretz <kretz@kde.org>