CTest: Improve CDash sorting and presentation
* Present both the git hash and branch name in the build name. The
former is useful for CI submissions, the latter is useful to more easily
identify the submission without looking at the hash.
* Drop OS info from the build name and rather attach it as a note. This
was too much noise in the build name.
* Use the Vc version string to identify the build group. Consequently
drop it from the build name where it would be redundant.
* Compile info on target architecture family and ABI used (e.g. x86 x32)
and use it for the build group. This is the most useful split of
submissions since 64-bit has high priority and 32 and x32 are rather
unimportant. Consequently drop the ${chip} info from the build name.

Signed-off-by: Matthias Kretz <kretz@kde.org>