Support broadcast conversion from non-const T
This is a corner case: If T is convertible to value_type, but `const T`
is not, and the argument to the constructor is non-const, then the
broadcast should still work. If the argument is `const U &x` then will
always be const in the context of the static_cast. At least Clang 3.8
thus failed to compile the unit test (`convertible` not convertible to
`int`, it said). Now the ctor uses a perfect forward and it works on
3.8, too.

Signed-off-by: Matthias Kretz <kretz@kde.org>