Don't use pext for now
I came here because the Intel SDE incorrectly modifies eflags when pext
is used. This randomly makes tests running in the SDE fail.

While looking at a workaround I realized that the 3 cycle latency of
pext is probably not the best solution for the bitset conversion. So go
with movemask + packs (for 16-bit) for now.

Signed-off-by: Matthias Kretz <kretz@kde.org>