SSE: Use MMX code only when necessary
The mask<float, sse> -> bool[4] conversion used _mm_cvtps_pi8 up to this
change. This conversion is not a single instruction but a sequence of
cvt ps pi32, and subsequent pack instructions. The SSE2 implementation
is more efficient. So when SSE2 is available, always disable the MMX
implementation.

Nice side effect: it avoids a bug in the Intel SDE, which fails to
execute the MMX pack instruction correctly.

Signed-off-by: Matthias Kretz <kretz@kde.org>