Split: avoid creation of uninitialized datapar objects
The old split implementation used subscripting to assign by element. For
subscript assignment to work it first created the array<V, Parts>, which
now (correctly) keeps the memory uninitialized. Due to the type-punning
nature of subscripting, this implied UB on assignment, then.

Now it uses the generator constructor to create a fully initialized
array<V, Parts> in a single pass.

Signed-off-by: Matthias Kretz <kretz@kde.org>