MSVC: once more, constexpr functions cannot be template args
This was close to impossible to find. No indication at the place that
needed a workaround, and lots of unclear error messages. In the end,
MSVC was just unable to produce the return type from the deduced
function template arguments, because the return type used a constexpr
function in its construction.

Signed-off-by: Matthias Kretz <kretz@kde.org>