Remove alignment workaround that doesn't work
alignof(double) on -m32 is 4. This hack tried to at least make
EntryType/value_type have the right alignment. But GCC drops attributes
in many situations (and warns about it when it does), leading to
warnings and failures. The static_assert safeguard in the gather/scatter
code suffices to make real problems ill-formed. So drop this useless
alignment attribute.

Signed-off-by: Matthias Kretz <kretz@kde.org>