Detect RandomAccessIterator to optimize load
Up to now a `simdize`d iterator would load a vector via iteration over
the elements and scalar iterator dereference + increment. Sadly, the
compiler does not optimize this into a vector load for std::vector
iterators. Now, the implementation dispatches on
Traits::has_contiguous_storage<It> to use an unaligned load instead of
scalar loads.

Detecting aligned storage is not possible, but it might be possible to
extend the API to communicate alignment via the simdized iterator type.

Fixes: gh-218

Signed-off-by: Matthias Kretz <kretz@kde.org>