Simplify integration of Vc into a parent project using CMake
This patch adds a Vc_submodule.cmake which can be included by
a CMake parent project. This builds the Vc target and sets the
corresponding compiler flags on Vc. The flags are also published
through the target to all consumers of it. So basically all you
need to do now is:

```
include(Vc/Vc_submodule.cmake)

add_executable(foo ...)
target_link_libraries(foo Vc)
```

The submodule file tries not to pollute the global CMake properties.
I.e. CMAKE_CXX_FLAGS is reset to the previous value. CMAKE_BUILD_TYPE
is also not touched by default - if that is preferred, consumers need
to `set(USE_VC_BUILDTYPE_FLAGS ON)`.

This patch fixes #188 for the 1.4 branch.