Allow building as subproject
When building Vc as a subproject of another CMake project, either
by including it as a git submodule or by using CMake's FetchContent,
the CMAKE_SOURCE_DIR variable does not point to the repository root.
Similary, the CMAKE_BINARY_DIR does no longer correspond to the
CMAKE_CURRENT_BINARY_DIR of the the top-level CMakeLists.txt file.

Using PROJECT_SOURCE_DIR and PROJECT_BINARY_DIR instead guarantees
that additional files (e.g. the CMake Config template or custom modules)
are found correctly and that no files of the consuming project in
the CMAKE_BINARY_DIR are overwritten.